+++ modules/platforms/cpp/core/src/impl/interop/interop_target.cpp
@@ -105,7 +105,7 @@ namespace ignite
 
                 if (outPtr)
                 {
-                    long long res = env.Get()->Context()->TargetInStreamOutLong(javaRef, opType, outPtr, &jniErr);
+                    int64_t res = env.Get()->Context()->TargetInStreamOutLong(javaRef, opType, outPtr, &jniErr);
 
                     IgniteError::SetError(jniErr.code, jniErr.errCls, jniErr.errMsg, err);
 
@@ -120,7 +120,7 @@ namespace ignite
             {
                 JniErrorInfo jniErr;
 
-                long long res = env.Get()->Context()->TargetInLongOutLong(javaRef, opType, 0, &jniErr);
+                int64_t res = env.Get()->Context()->TargetInLongOutLong(javaRef, opType, 0, &jniErr);
 
                 IgniteError::SetError(jniErr.code, jniErr.errCls, jniErr.errMsg, err);
 
@@ -248,7 +248,7 @@ namespace ignite
             {
                 JniErrorInfo jniErr;
 
-                long long res = env.Get()->Context()->TargetInLongOutLong(javaRef, opType, val, &jniErr);
+                int64_t res = env.Get()->Context()->TargetInLongOutLong(javaRef, opType, val, &jniErr);
 
                 IgniteError::SetError(jniErr.code, jniErr.errCls, jniErr.errMsg, err);
